Regression at single trial level
Including regression at the single trial level, as defined in the design
for each condition. If specifying a SPM.mat as design, need to go back
to 'specify design' and enter the targets, which then take the block
structure (potential replication of targets) and discarded trials/scans
into account. New window (similar to classification) to select the
regression target, and modified the compute_targets in the prt_model
code. It needs much more testing!