Allows the output of the PRT struct instead of a filename.
For prt_cv_model:
This is just to allow easier scripting: sometimes it's useful/easier if
the output of the CV is the PRT structure itself, instead of the
filename of where it's saved. The dafault behaviour is preserved to
ensure back compatibility.
For prt_model:
This is just to decide if the PRT should be saved at the end or not. it
is anyway returned as an output argument.