Fixed bug - Error from undefined hrt_image name.
If I want to recompute the weights, there is an if-clause in #~258 (in *_class.m) that checks if the files of the weights exist, and if so it deletes them first.

In case of .mat files there is no hdr image, only the .mat file itself, so the hdr_image isn’t assigned a name, and then when it tries to delete(hdr_image) it throws an error because it is undefined. Hence the change.