Quick bug fix when computing and plotting weights
This bug came up when trying to compute and plot weights in a KRR model where we had concatenated (imaging) .nifti with behavioural (.mat) modalities.

Further investigation might be needed with regards to parsing of the atlas string in GUI (in structure within the prt_ui_compute_weights) & Batch, as well as the use of the "updated_aal_atlas" file within PRoNTo.